refactor: split the frontend JS into ES modules built to assets/build
Breaks the single-file IIFE into concern-scoped modules: registries
(public extension globals + template contract docs), config, escape,
the built-in table template, the core default renderers/request/focus,
and the per-instance controller. The enqueue now loads the bundled
assets/build/meeting-minutes.js using the generated .asset.php for the
cache-busting version. No behaviour change: the window globals
(edmmExtraColumns, edmmTemplates, edmmEscapeAttr), template contract,
markup, and REST query are identical.
